Add files via upload
New Assembly is the Assembly with the design. I have hit a bit of a wall when it comes to stuff to add, and how to make the Y axis move. It is still pretty rough as far as it goes, but all of the parts are at least close to what they will be in the final design. The floating motor will be attached to the case using a bracket, or the frame if we can figure out a way to do that.